First working transaction processing!
In AssetFrame, changed to use start a TransactionFrame vice a TransactionGrid (Grid will be embedded in the Frame).

In AssetGrid, changed OnCellLeftClick to start a TransactionFrame vice a TransactionGrid.

Changed TransactionFrame to behave similarly to AssetFrame.

Changed TransactionGrid to look more like AssetGrid.

Idea is to refactor commonalities out of AssetFrame and TransactionFrame as well as AssetGrid and TransactionGrid at a later time!